Add mesh

+++ MagicWorld/VertexBuffer.h
@@ -6,7 +6,7 @@
 class VertexBuffer
 {
 public:
-	VertexBuffer(const void* data, unsigned int size);
+	VertexBuffer(const void* data, unsigned long long size);
 	~VertexBuffer();
 
 	VertexBuffer(VertexBuffer&& other) noexcept;
